[Feat]: UseCase 기본 interface 추가
