+++ src/main/java/team/haedal/gifticionfunding/service/common/UseCase.java
@@ -1,0 +1,5 @@
+package team.haedal.gifticionfunding.service.common;
+
+public interface UseCase <Input, Output>{
+    Output invoke(Input input);
+}
